requirements.txt

fastapi
uvicorn[standard]
caldav
icalendar
google-auth
google-auth-oauthlib
requests
python-dotenv
python3-uvicorn
cachetools